Check in correct versions of files.  Somehow null files got committed last time.
There's a possibility that there are still things wrong with the
commit.  I wasn't able to build the version I committed, and I won't
be for another few days.  If any committer finds further omissions,
the original source archive is on freefall:~grog/picobsd.tar.gz.  Feel
free to correct.

Reported-by: "Louis A. Mamakos" <louie@TransSys.COM>


git-svn-id: svn+ssh://svn.freebsd.org/base/head@54437 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f