The USB double bulk pipe driver (Host to host cables). Currently there are two supported chips, the NetChip 1080 (only prototypes available) and the EzLink cable. Any other cable should be supported however as they are all very much alike (there is a difference between them wrt performance).
It uses Netgraph.

This driver was mostly written by Doug Ambrisko and Julian Elischer and
I would like to thank Whistle for yet another contribution. And my
aplogies to them for me sitting on the driver for so long (2 months).

Also, many thanks to Reid Augustin from NetChip for providing me with a
prototype of their 1080 chip.

Be aware of the fact that this driver is very immature and has only been
tested very lightly. If someone feels like learning about Netgraph however
this is an excellent driver to start playing with.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@59877 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f