* Bring back the guts of crt{i,n}.S.  This allows C++ exceptions to work when using the egcs and gcc-devel ports, along with GCC built from stock public FSF sources.  With out this change, FreeBSD will be removed from the list of systems GCC 3.0 must be evaluated on before release.  With the effort some of us put into getting FreeBSD on this list, we should not turn this effort into a waste, else we might not be worth fighting for in the future.  (note that Alpha and IA-64 versions of crt{i,n}.S are needed)
* Switch from our own crt{begin,in} to those created from GCC's crtstuff.c.
This will allow us to switch to DWARF2 exceptions in the future, along with
staying in sync with any future GCC requirements.

* Break out our ELF branding bits into a seperate file.  Currently this
is now included by our crt1.c files (since this functionality was part of
our native crtbegin.c).  Later crtbrand.o will be merged in the creation
of crti.o.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@67811 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f