Switch to using a struct xucred instead of a struct xucred when not actually in the kernel.  This structure is a different size than what is currently in -CURRENT, but should hopefully be the last time any application breakage is caused there.  As soon as any major inconveniences are removed, the definition of the in-kernel struct ucred should be conditionalized upon defined(_KERNEL).
This also changes struct export_args to remove dependency on the
constantly-changing struct ucred, as well as limiting the bounds
of the size fields to the correct size.  This means: a) mountd and
friends won't break all the time, b) mountd and friends won't crash
the kernel all the time if they don't know what they're doing wrt
actual struct export_args layout.

Reviewed by:	bde


git-svn-id: svn+ssh://svn.freebsd.org/base/head@72650 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f