Fix the handling of VJ uncompression.  Unfortunately, tcp_uncompress() makes the implied assumption there were another 128 bytes of space in front of the packet handed off to it... which is not the case for sppp.  This could easily end up in corrupting random memory.
This fix is about the same as revs 1.6, 1.8, and 1.9 from our
i4b_ispppsubr.c.

Also fixed IPCP option negotiation to zero out the options when
starting IPCP.  Otherwise, if negotiation parameters change between
various IPCP startups, it could happen that old options would still be
requested (this happened if VJ was turned off, and ended up in half
off the link still negotiating for VJ compression).

IMHO, the base system's sppp is now feature-wise up to date with the
one in the i4b part of the tree, so the latter can be disabled.

MFC after:	1 month


git-svn-id: svn+ssh://svn.freebsd.org/base/head@88700 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f