Add hooks for very basic IPFilter support in bridging. Set,
  # sysctl net.link.ether.bdg_ipf=1

To enable. Just like ipfw(8) bridging, only input packets are filtered
in the bridge. Filtering works just like in the IP layer, ipf(8)
first, then ipfw(8). And just like in the IP layer, both are
independent, one need not be run to use the other. (Note: This will
not work in, but doesn't break, the bridge.ko module. The ipl.ko
module would need to be fixed before that is worth worrying about.)

Reviewed by:	luigi


git-svn-id: svn+ssh://svn.freebsd.org/base/head@92522 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f