Add support for an additional field to the packages/INDEX file.  If present, this field specifies the media volume that the disc is contained on.  If the volume of a given packages is different than the current volume of mediaDevice, then the user is prompted --
  "This is disc #%d.  Package %s is on disc #%d\n"
  "Would you like to switch discs now?\n"

If the user selects yes, then DEVICE_SHUTDOWN is called and the user
is then prompted --

  "Please remove disc #%d from you drive, and add disc #%d"

This works well for a carefully crafted INDEX file, but more work
needs to be done to sort dependencies on a given package based on the
volume that they reside on, to minimize the amount of disc flipping
required of the user.

This commit is a no-op for normal INDEX files and FreeBSD CDs.  These
additional features are only used if the INDEX and cdrom.inf file have
multi-volume support.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@94601 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f