Make the consumers of the linker_load_file() function use linker_load_module() instead.
This fixes a bug where the kernel was unable to properly locate and
load a kernel module in vfs_mount() (and probably in the netgraph
code as well since it was using the same function).  This is because
the linker_load_file() does not properly search the module path.

Problem found by:	peter
Reviewed by:		peter
Thanks to:		peter


git-svn-id: svn+ssh://svn.freebsd.org/base/head@101241 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f