BANG! BANG! BANG!  Put these bits out of their misery. Murdered by members of:	a.out.die.die.die
ELF is the 1 true path now.  So make good on the src/Makefile threat that
building a.out will not be supported post 4.x.  These bits should either
resurface as a port, or a new port using the latest Binutils bits.  The
later will not support our SunOS-style shared a.out libs; but we shouldn't
need such support by this point in time.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@103437 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f